tests/wpt/tests/html/canvas/offscreen/canvas-context/2d.canvas.context.type.replace.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.context.type.replace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.context.type.replace</h1>
<p class="desc">Interface methods can be overridden</p>

<p class="notes">Defined in "Web IDL" (draft)
<script>
var t = async_test("Interface methods can be overridden");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var fillRect = window.OffscreenCanvasRenderingContext2D.prototype.fillRect;
  window.OffscreenCanvasRenderingContext2D.prototype.fillRect = function (x, y, w, h)
  {
      this.fillStyle = '#0f0';
      fillRect.call(this, x, y, w, h);
  };
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
